LayoutTests/imported/mozilla/css-animations/test_keyframeeffect-getkeyframes-expected.txt


FAIL KeyframeEffectReadOnly.getKeyframes() returns no frames for various kinds of empty enimations assert_equals: number of frames when @keyframes has empty keyframes expected 0 but got 2
PASS KeyframeEffectReadOnly.getKeyframes() returns expected frames for a simple animation 
PASS KeyframeEffectReadOnly.getKeyframes() returns frames with expected easing values, when the easing comes from animation-timing-function on the element 
FAIL KeyframeEffectReadOnly.getKeyframes() returns frames with expected easing values, when the easing is specified on each keyframe assert_equals: value of 'easing' on ComputedKeyframe #1 expected "ease-in-out" but got "linear"
FAIL KeyframeEffectReadOnly.getKeyframes() returns frames with expected easing values, when the easing is specified on some keyframes assert_equals: value of 'easing' on ComputedKeyframe #1 expected "steps(1, start)" but got "linear"
PASS KeyframeEffectReadOnly.getKeyframes() returns expected frames for a simple animation that specifies a single shorthand property 
FAIL KeyframeEffectReadOnly.getKeyframes() returns expected frames for an animation with a 0% keyframe and no 100% keyframe assert_equals: properties on ComputedKeyframe #1 expected "color,composite,computedOffset,easing,offset" but got "composite,computedOffset,easing,offset"
FAIL KeyframeEffectReadOnly.getKeyframes() returns expected frames for an animation with a 100% keyframe and no 0% keyframe assert_equals: properties on ComputedKeyframe #0 expected "color,composite,computedOffset,easing,offset" but got "composite,computedOffset,easing,offset"
FAIL KeyframeEffectReadOnly.getKeyframes() returns expected frames for an animation with no 0% or 100% keyframe but with a 50% keyframe assert_equals: properties on ComputedKeyframe #0 expected "color,composite,computedOffset,easing,offset" but got "composite,computedOffset,easing,offset"
FAIL KeyframeEffectReadOnly.getKeyframes() returns expected frames for an animation with a partially complete 100% keyframe (because the !important rule is ignored) assert_equals: properties on ComputedKeyframe #1 expected "composite,computedOffset,easing,marginBottom,marginTop,offset" but got "composite,computedOffset,easing,marginBottom,offset"
PASS KeyframeEffectReadOnly.getKeyframes() returns expected frames for an animation with different properties on different keyframes, all with the same easing function 
FAIL KeyframeEffectReadOnly.getKeyframes() returns expected frames for an animation with different properties on different keyframes, with a different easing function on each assert_equals: value for 'easing' on ComputedKeyframe #1 expected "steps(1)" but got "linear"
PASS KeyframeEffectReadOnly.getKeyframes() returns expected frames for an animation with multiple keyframes for the same time, and all with the same easing function 
FAIL KeyframeEffectReadOnly.getKeyframes() returns expected frames for an animation with multiple keyframes for the same time and with different easing functions assert_equals: number of frames expected 3 but got 2
FAIL KeyframeEffectReadOnly.getKeyframes() returns expected frames for an animation with multiple keyframes for the same time and with different but equivalent easing functions assert_equals: value for 'easing' on ComputedKeyframe #2 expected "ease" but got "steps(1)"
PASS KeyframeEffectReadOnly.getKeyframes() returns expected frames for overlapping keyframes 
FAIL KeyframeEffectReadOnly.getKeyframes() returns expected values for animations with filter properties and missing keyframes assert_equals: properties on ComputedKeyframe #0 expected "composite,computedOffset,easing,filter,offset" but got "composite,computedOffset,easing,offset"
PASS KeyframeEffectReadOnly.getKeyframes() returns expected values for animation with drop-shadow of filter property 
FAIL KeyframeEffectReadOnly.getKeyframes() returns expected values for animations with text-shadow properties and missing keyframes assert_equals: properties on ComputedKeyframe #0 expected "composite,computedOffset,easing,offset,textShadow" but got "composite,computedOffset,easing,offset"
FAIL KeyframeEffectReadOnly.getKeyframes() returns expected values for animations with background-size properties and missing keyframes assert_equals: properties on ComputedKeyframe #0 expected "backgroundSize,composite,computedOffset,easing,offset" but got "composite,computedOffset,easing,offset"
FAIL KeyframeEffectReadOnly.getKeyframes() returns expected values for animations with CSS variables as keyframe values assert_equals: properties on ComputedKeyframe #0 expected "composite,computedOffset,easing,offset,transform" but got "composite,computedOffset,easing,offset"
FAIL KeyframeEffectReadOnly.getKeyframes() returns expected values for animations with CSS variables as keyframe values in a shorthand property assert_equals: properties on ComputedKeyframe #0 expected "composite,computedOffset,easing,marginBottom,marginLeft,marginRight,marginTop,offset" but got "composite,computedOffset,easing,offset"
FAIL KeyframeEffectReadOnly.getKeyframes() returns expected values for animations with a CSS variable which is overriden by the value in keyframe assert_equals: properties on ComputedKeyframe #0 expected "color,composite,computedOffset,easing,offset" but got "composite,computedOffset,easing,offset"
FAIL KeyframeEffectReadOnly.getKeyframes() returns expected values for animations with only custom property in a keyframe assert_equals: value for 'transform' on ComputedKeyframe #0 expected "translate(100px, 0px)" but got "none"
